Automatically link to sources
